Restructured template generator module.
