+++ client/src/app/template-generator/questionnaire/template-questionnaire.component.ts
@@ -1,5 +1,5 @@
 import { Component, OnInit } from '@angular/core';
-import {TemplateSpecification} from "./template-specification";
+import {TemplateSpecification} from "../template-specification";
 
 @Component({
   selector: 'app-template-questionnaire',
